+++ src/idpconfgen/components/bgeo_strategies/__init__.py
@@ -1,16 +1,16 @@
 """Bond geomtry strategies."""
 import argparse
 
+from idpconfgen.components.bgeo_strategies.fixed import name as bgeo_fixed_name
 from idpconfgen.components.bgeo_strategies.int2cart.bgeo_int2cart import (
     has_int2cart,
     )
 from idpconfgen.components.bgeo_strategies.int2cart.bgeo_int2cart import \
     name as bgeo_int2cart_name
-from idpconfgen.components.bgeo_strategies.fixed import \
-    name as bgeo_fixed_name
 from idpconfgen.components.bgeo_strategies.sampling import \
     name as bgeo_sampling_name
-
+from idpconfgen.components.bgeo_strategies.exact import \
+    name as bgeo_exact_name
 
 bgeo_strategies_default = bgeo_sampling_name
 """The default bond geometry sampling strategy."""
@@ -18,6 +18,7 @@ bgeo_strategies_default = bgeo_sampling_name
 bgeo_strategies = (
     bgeo_fixed_name,
     bgeo_sampling_name,
+    bgeo_exact_name,
     bgeo_int2cart_name,
     )
 """Available bond geometry sampling strategies."""
@@ -30,7 +31,10 @@ def add_bgeo_strategy_arg(parse):
     parse.add_argument(
         _bgeo_arg_name,
         dest='bgeo_strategy',
-        help="Which strategy to use for bond geometries. Defaults to `sampling`.",
+        help=(
+            "Which strategy to use for bond geometries. "
+            "Defaults to `sampling`."
+            ),
         choices=bgeo_strategies,
         default=bgeo_strategies_default,
         action=CheckBgeoInstallation,