+++ src/idpconfgen/components/bgeo_strategies/int2cart/bgeo_int2cart.py
@@ -2,10 +2,10 @@
 import os
 from pathlib import Path
 
+
 try:
     import torch
     import yaml
-
     from modelling.models.builder import BackboneBuilder
     from modelling.utils.predict import predict
     has_int2cart = True